Various ui fixes
